requirements.txt
google-cloud-vision==3.7.2
python-dotenv==1.0.1
pytest==8.3.4
send2trash==1.8.3
